+++ src/init.c
@@ -35,7 +35,14 @@ extern SEXP serialize_df_to_ndjson_file_(SEXP robj_, SEXP filename_, SEXP serial
 extern SEXP serialize_list_to_ndjson_str_ (SEXP robj_,                 SEXP serialize_opts_);
 extern SEXP serialize_list_to_ndjson_file_(SEXP robj_, SEXP filename_, SEXP serialize_opts_);
 
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// GeoJSON
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+extern SEXP parse_geojson_str_ (SEXP str_     , SEXP geo_opts_, SEXP parse_opts_);
+extern SEXP parse_geojson_file_(SEXP filename_, SEXP geo_opts_, SEXP parse_opts_);
 
+extern SEXP serialize_sf_to_str_ (SEXP sf_                , SEXP geo_opts_, SEXP serialize_opts_);
+extern SEXP serialize_sf_to_file_(SEXP sf_, SEXP filename_, SEXP geo_opts_, SEXP serialize_opts_);
 
 
 static const R_CallMethodDef CEntries[] = {
@@ -69,6 +76,15 @@ static const R_CallMethodDef CEntries[] = {
   
   {"serialize_list_to_ndjson_str_" , (DL_FUNC) &serialize_list_to_ndjson_str_ , 2},
   {"serialize_list_to_ndjson_file_", (DL_FUNC) &serialize_list_to_ndjson_file_, 3},
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // GeoJSON
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  {"parse_geojson_str_" , (DL_FUNC) &parse_geojson_str_ , 3},
+  {"parse_geojson_file_", (DL_FUNC) &parse_geojson_file_, 3},
+
+  {"serialize_sf_to_str_" , (DL_FUNC) &serialize_sf_to_str_ , 3},
+  {"serialize_sf_to_file_", (DL_FUNC) &serialize_sf_to_file_, 4},
   
   
   {NULL , NULL, 0}